areclambord_api/requirements.txt
asgiref==3.2.7
bson==0.5.8
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.2.13
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
jws==0.1.3
MarkupSafe==1.1.1
numpy==1.19.0
oauth2client==4.1.3
openapi-codec==1.3.2
opencv-python==4.2.0.34
Pillow==7.2.0
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.6
pymongo==3.10.1
pyparsing==2.4.7
Pyrebase-t==1.0.28
Pyrebase4==4.4.3
pytesseract==0.3.4
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
